Notebook to generate synthetic magnetic data
Generated from 3 bodies: a dike, a sill and a batholith.
Bodies are made from PolygonalPrisms. The vertices of each are in the
corresponding .json files. model.pickle is a list of the PolygonalPrism
instances. metadata.json has the grid shape, area, mesh bounds,
inclination and declination. synthetic_data.txt is x, y, z, total field.